Use "digits" instead of "decimals" in the README
This changes the word "decimal" to "fractional" when talking about a number with a fractional part and the word "decimal" to "digit" when talking about the digits of the fractional part of that number.  Decimal usually means base-10, and using the word "decimal" for both binary and decimal digits is a bit confusing.